requirements.txt
pandas
sqlalchemy
psycopg2-binary
tqdm
streamlit
python-dotenv